migrated the share logic
